+++ src/lib.rs
@@ -74,7 +74,9 @@ pub use bitvec;
 pub use compress::{CompressOptions, DitherMethod};
 #[cfg(feature = "ndarray")]
 pub use data::ImageArray;
-pub use data::{Image, ImageData, ImageView, RawImage, SampleType, Scaling, UnsignedView};
+pub use data::{
+    Image, ImageData, ImageMetadata, ImageView, RawImage, SampleType, Scaling, UnsignedView,
+};
 pub use error::{FitsError, Result};
 pub use groups::{RandomGroupView, RandomGroups};
 pub use hdu::HduKind;